+++ src/main/java/net/grid/vampiresdelight/common/VDFoodValues.java
@@ -142,22 +142,22 @@ public class VDFoodValues {
             .nutrition(3).saturationMod(0.2f).fast()
             .effect(() -> new MobEffectInstance(MobEffects.MOVEMENT_SPEED, FLEETING_DURATION), 1.0F).build();
     public static final FoodProperties SNOW_WHITE_ICE_CREAM = (new FoodProperties.Builder())
-            .nutrition(7).saturationMod(0.6f).fast()
+            .nutrition(8).saturationMod(0.6f).fast()
             .effect(() -> new MobEffectInstance(VDEffects.BLESSING.get(), MEDIUM_DURATION), 1.0F).build();
 
     // Handheld Foods
     public static final FoodProperties RICE_BREAD = (new FoodProperties.Builder())
             .nutrition(5).saturationMod(0.6F).build();
     public static final FoodProperties FISH_BURGER = (new FoodProperties.Builder())
-            .nutrition(12).saturationMod(0.9f).build();
+            .nutrition(12).saturationMod(0.8f).build();
     public static final FoodProperties BLOOD_SAUSAGE = (new FoodProperties.Builder())
             .nutrition(10).saturationMod(0.8f).build();
     public static final FoodProperties BLOOD_HOT_DOG = (new FoodProperties.Builder())
-            .nutrition(14).saturationMod(0.9f).build();
+            .nutrition(14).saturationMod(0.8f).build();
     public static final FoodProperties BLOOD_BAGEL = (new FoodProperties.Builder())
-            .nutrition(9).saturationMod(0.8f).build();
+            .nutrition(9).saturationMod(0.7f).build();
     public static final FoodProperties BAGEL_SANDWICH = (new FoodProperties.Builder())
-            .nutrition(14).saturationMod(0.95f).build();
+            .nutrition(14).saturationMod(0.8f).build();
     public static final FoodProperties EYES_ON_STICK = (new FoodProperties.Builder())
             .nutrition(7).saturationMod(0.6f).build();
     public static final FoodProperties EYE_CROISSANT = (new FoodProperties.Builder())
@@ -203,9 +203,9 @@ public class VDFoodValues {
             .nutrition(7).saturationMod(0.8f).fast().alwaysEat()
             .effect(() -> {
                 if (VDConfiguration.REPLACE_WEIRD_JELLY_SUNSCREEN_WITH_JUMPBOOST.get()) {
-                    return new MobEffectInstance(MobEffects.JUMP, BRIEF_DURATION);
+                    return new MobEffectInstance(MobEffects.JUMP, SHORT_DURATION);
                 } else {
-                    return new MobEffectInstance(de.teamlapen.vampirism.core.ModEffects.SUNSCREEN.get(), BRIEF_DURATION);
+                    return new MobEffectInstance(de.teamlapen.vampirism.core.ModEffects.SUNSCREEN.get(), SHORT_DURATION);
                 }
             }, 1.0F).build();